fix(mypy): clear free-zone judgement errors + 2 real bugs (Item D batch 3-5)
Drives the own-file mypy errors in features/signals/portfolio from 60 to 0
(stash-confirmed; the 11 remaining errors are all imported from data/+execution/,
their own batches). Part of the staged mypy 134->0 effort. No signal/weight
regression (senior + test-runner verified; multifactor IC-path narrowing is
assert-only).

TYPE-ONLY (byte-identical — precise annotations / value-preserving float() casts on
already-numeric scalars / typed ndarray locals mypy widened to Any / runtime-no-op
narrowings):
- no-any-return: seasonal_features, tsfresh_augmentation, chart_pattern_matrix,
  lppls_crash, covariance, hierarchical_risk_parity, optimizers, black_litterman,
  news_features, conformal_position, inverse_etf_selector, tail_risk_hedge,
  crash_prediction, analyst_revisions.
- attr-defined on untyped-optional-dep params -> Any: macro_regime_quadrant,
  sentiment_panel, recession_probability (fred_client), dro_portfolio, position_sizing
  (entity_linker).
- arg-type / dict-item / assignment / var-annotated / return-value: covariance
  (Literal dcc method), optimizers, stress_test_constraints, quantum_portfolio,
  multiasset_allocator (bool(np.bool_)), ta_factors_core, triple_barrier (rename),
  event_features, factor_store_integration (Timestamp locals), rules_trend.
- conformal_position._mapie: Any | None (reassigned before use; other sites guarded).
- multifactor_signal:628/656: assert ic_weights_df is not None inside `if use_ic:`
  (use_ic implies non-None) -> pure narrowing, no weight/mf_score change.
- meta_model: meta_labeler: Any | None + removed the wrong type: ignore[union-attr];
  cast on safe_load_model return (SHA256-verified, no runtime effect).

REAL-BUGS (genuinely fixed, NOT silenced, each tested):
1. signals/meta_model.py explain_prediction_lime referenced LimeExplanation
   attributes top_features()/predicted_value that DO NOT EXIST (the dataclass has
   only feature_contributions + source). The call always raised AttributeError,
   swallowed by the surrounding try/except -> the function ALWAYS returned
   {"error": ...} (explainability never worked). Fixed: coerce training_data/X_row
   to ndarray aligned to feature_names, derive top_features from feature_contributions
   sorted by |contribution|, return the real schema, drop the fabricated
   predicted_value. No external caller used the old keys. + tests/test_meta_model_lime_explanation.py (3 tests).
2. signals/news_validation.py:217 beta (float|None) was passed to
   compute_abnormal_returns(beta: float) with only an alpha-None guard. Since
   compute_market_model returns alpha/beta strictly all-None-or-all-float, added
   `or beta is None` — behaviour-identical + type-correct. + regression test.

ruff clean. 385 passed / 0 failed / 25 skipped (pre-existing optional-dep gates) on
the touched + new suites. senior PASS (type_only_byte_identical: yes,
real_bugs_correctly_fixed: yes), test-runner PASS (own_file_errors_cleared: yes,
no_signal_regression: yes, newly_introduced_failures: []), auditor PASS. CI not run.

New tests: tests/test_meta_model_lime_explanation.py, tests/test_multifactor_adaptive_ic_narrowing.py;
modified: tests/test_news_validation.py.

+++ src/assembled_core/portfolio/inverse_etf_selector.py
@@ -273,7 +273,7 @@ class InverseETFSelector:
         # Gross leveraged return
         daily_underlying = annualized_return / 252
         leveraged_daily = leverage * daily_underlying
-        net_return = leveraged_daily * holding_days - decay
+        net_return = float(leveraged_daily * holding_days - decay)
 
         logger.debug(
             "[InverseETFSelector] %s decay=%.3f gross_return=%.3f net=%.3f (hold=%dd)",
